// src/app/models/admin/capphephoatdongkhoangsan/cpkhaithackhoangsan/cpkhaithacdvhc.model.ts
export class OutputCpKhaiThacDVHCModel {
    idkhaithacdvhc: string;
    idxa: string;
    idhuyen: string;
    idtinh: string;
    maxa: string;
    mahuyen: string;
    matinh: string;
    tentinh: string;
    tenhuyen: string;
    tenxa: string;
    idcapphepkhaithac: string;
    siteid: string;
}

export class InputCpKhaiThacDVHCModel {
    idkhaithacdvhc: string;
    idxa: string;
    idhuyen: string;
    idtinh: string;
    maxa: string;
    mahuyen: string;
    matinh: string;
    idcapphepkhaithac: string;
}